Fix React Oauth issue
